Method_4/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.1
joblib>=1.3.2
customtkinter>=5.2.0
pillow>=10.0.0
matplotlib>=3.7.1
